requirements.txt
sentence-transformers
faiss-cpu
pypdf
ollama
